requirements.txt
telethon==1.36.0
pymongo==4.6.1
dnspython==2.4.2
python-dotenv==1.0.0
flask==3.0.0
werkzeug==3.0.1
